Improve package naming concistency
The method naming conventions is more inline with PHP naming convention.

The Collection Interface and the classes that implemetns it are updated
accordingly. The "offset" keyword is replaced by the "key" keyword.

The Host::__toString now returns the string representation formatted
as is was submitted. the Host::toAscii and Host::isIdn methods are
added to help the developer. the IPv6 zone identifer manipulation API
is completed with Host::hasZoneIdentifer method