Remove the AbstractUri class
To simplify the URI object extension and creation the AbstractUri
abstract class is removed in favor of a GenericUriTrait. This way
the Path Interface is respected and extended URI object can still
use an extend version of the Path Interface.
The readonly property are no longer attached to the interface but
to the concrete classes.